Move `Prompt` to the `arg` package
